requirements.txt
fastapi
uvicorn
lyricsgenius
openai
python-dotenv
beautifulsoup4
requests
